- expanded lint script to ensure deterministic ordering of more things
