+++ src/buttonwood-bonds.tokenlist.json
@@ -1,6 +1,6 @@
 {
   "name": "Buttonwood Bonds",
-  "timestamp": "2023-08-09T14:43:05.382Z",
+  "timestamp": "2023-08-09T14:58:46.605Z",
   "version": {
     "major": 0,
     "minor": 4,
