+++ src/buttonwood.tokenlist.json
@@ -1,6 +1,6 @@
 {
   "name": "Buttonwood",
-  "timestamp": "2023-08-09T14:43:05.379Z",
+  "timestamp": "2023-08-09T14:58:46.602Z",
   "version": {
     "major": 0,
     "minor": 4,
