+++ src/tokens.json
@@ -56,11 +56,11 @@
     "decimals": 6,
     "primaryChain": "Avalanche C-Chain",
     "chains": {
-      "Avalanche C-Chain": {
-        "address": "0x625E7708f30cA75bfd92586e17077590C60eb4cD"
-      },
       "Avalanche Fuji Testnet": {
         "address": "0x2c4a078f1FC5B545f3103c870d22f9AC5F0F673E"
+      },
+      "Avalanche C-Chain": {
+        "address": "0x625E7708f30cA75bfd92586e17077590C60eb4cD"
       }
     },
     "tags": [
@@ -74,11 +74,11 @@
     "decimals": 6,
     "primaryChain": "Avalanche C-Chain",
     "chains": {
-      "Avalanche C-Chain": {
-        "address": "0x6ab707Aca953eDAeFBc4fD23bA73294241490620"
-      },
       "Avalanche Fuji Testnet": {
         "address": "0x6392E5A601620aC1E28747f3428861dc4562CFfA"
+      },
+      "Avalanche C-Chain": {
+        "address": "0x6ab707Aca953eDAeFBc4fD23bA73294241490620"
       }
     },
     "tags": [
@@ -95,11 +95,11 @@
       "Ethereum Mainnet": {
         "address": "0x98C23E9d8f34FEFb1B7BD6a91B7FF122F4e16F5c"
       },
-      "Sepolia": {
-        "address": "0x55D45c6649a0Ff74097d66aa6A6ae18a66Bb2fF3"
-      },
       "Goerli": {
         "address": "0x8Be59D90A7Dc679C5cE5a7963cD1082dAB499918"
+      },
+      "Sepolia": {
+        "address": "0x55D45c6649a0Ff74097d66aa6A6ae18a66Bb2fF3"
       }
     },
     "tags": [
@@ -116,11 +116,11 @@
       "Ethereum Mainnet": {
         "address": "0x23878914EFE38d27C4D67Ab83ed1b93A74D4086a"
       },
-      "Sepolia": {
-        "address": "0xFbE6E10f1E7B15e2e7904a5ca249a8b6dF8d4309"
-      },
       "Goerli": {
         "address": "0xf3368D1383cE079006E5D1d56878b92bbf08F1c2"
+      },
+      "Sepolia": {
+        "address": "0xFbE6E10f1E7B15e2e7904a5ca249a8b6dF8d4309"
       }
     },
     "tags": [
@@ -276,14 +276,14 @@
       "Ethereum Mainnet": {
         "address": "0xc3d688B66703497DAA19211EEdff47f25384cdc3"
       },
-      "Avalanche Fuji Testnet": {
-        "address": "0x59BF4753899C20EA152dEefc6f6A14B2a5CC3021"
-      },
       "Goerli": {
         "address": "0x3EE77595A8459e93C2888b13aDB354017B198188"
       },
       "Arbitrum One": {
         "address": "0xA5EDBDD9646f8dFF606d7448e414884C7d905dCA"
+      },
+      "Avalanche Fuji Testnet": {
+        "address": "0x59BF4753899C20EA152dEefc6f6A14B2a5CC3021"
       },
       "Arbitrum Goerli": {
         "address": "0x1d573274E19174260c5aCE3f2251598959d24456"
